refactor: extract CLI command handlers to commands.rs (1941→1728)
Move run_smoke_test, run_init, run_doctor, show_profile, run_help
to commands.rs. Creates 213 lines of headroom under the ratchet ceiling
for future provenance pipeline features.
